+++ buildSrc/src/main/kotlin/Globals.kt
@@ -1,15 +1,16 @@
 object Versions {
     // Misc and plugins
-    const val GRADLE = "6.8.2"
-    const val KOTLIN = "1.4.30"
+    const val GRADLE = "6.8.3"
+    const val KOTLIN = "1.4.31"
     const val BEN_MANES_VERSIONS_PLUGIN = "0.36.0"
     const val DOKKA_PLUGIN = "1.4.20"
 
     // Lib dependencies
-    const val KOTLIN_SERIALIZATION = "1.4.30"
-    const val KTOR = "1.5.1"
+    const val KOTLIN_SERIALIZATION = "1.4.31"
+    const val KTOR = "1.5.2"
     const val COROUTINES = "1.4.2"
     const val SLF4J = "1.7.30"
+    const val KOTLINX_CLI = "0.3.2"
 
     // Testing dependencies
     const val JUNIT = "4.13.1"