+++ library/src/commonMain/kotlin/org/jraf/klibnotion/model/property/value/rollup/ArrayRollupPropertyValue.kt
@@ -22,20 +22,14 @@
  * limitations under the License.
  */
 
-package org.jraf.klibnotion.internal.client
+package org.jraf.klibnotion.model.property.value.rollup
 
-import io.ktor.client.HttpClient
-import io.ktor.client.HttpClientConfig
-import io.ktor.client.engine.curl.Curl
+import org.jraf.klibnotion.model.property.value.PropertyValue
 
-internal actual fun createHttpClient(
-    bypassSslChecks: Boolean,
-    block: HttpClientConfig<*>.() -> Unit,
-): HttpClient {
-    return HttpClient(Curl) {
-        apply(block)
-        if (bypassSslChecks) engine {
-            sslVerify = false
-        }
-    }
-}
+/**
+ * See [https://www.notion.so/5a48631ae00c4d48adee859475a25956?v=5dfe884a62304ae08f1fb7d0e89c5743].
+ *
+ * Note: the [PropertyValue]s in the list will have empty string (`""`)
+ * for the [PropertyValue.name] and [PropertyValue.id] fields.
+ */
+interface ArrayRollupPropertyValue : RollupPropertyValue<List<PropertyValue<*>>>