+++ CHANGELOG.md
@@ -1,9 +1,11 @@
 # Changelog
 
-## vX.Y.Z (XXXX-XX-XX)
+## v1.1.0 (2021-02-27)
 
+- Support for Array Rollup property type
 - Slightly improve `DatabaseQuery` API (`any` instead of `addAnyFilters`, `all` instead of `addAllFilters`)
 - Slightly improve `DatabaseQuerySort` API (`ascending` and `descending` instead of `add`)
+- Add a `bypassSslChecks` option on `HttpConfiguration` (should never be used in production)
 
 ## v1.0.0 (2021-02-14)